Rename icon open-book into read
