+++ packages/core/src/icons/all.css
@@ -26,7 +26,7 @@
 @import "./icon-location.css";
 @import "./icon-lock.css";
 @import "./icon-menu.css";
-@import "./icon-open-book.css";
+@import "./icon-read.css";
 @import "./icon-search.css";
 @import "./icon-share.css";
 @import "./icon-smile.css";
