+++ packages/core/src/icons/icon-read.css
@@ -1,6 +1,6 @@
-@svg-load icon url(../svg/icon-open-book.svg);
+@svg-load icon url(../svg/icon-read.svg);
 
-.hds-icon--open-book {
+.hds-icon--read {
   mask-image: svg-inline(icon);
   -webkit-mask-image: svg-inline(icon);
 }
