+++ packages/react/src/icons/IconRead.tsx
